Clean up obsolete files for unit tests
