+++ libcaf_core/test/core-test.cpp
